Move getting vote gas limit to utils
